Correct typo in the Introduction
As a native English speaker I know this is wrong, but I lack the grammatical knowledge to explain why on a technical level. I resorted to using GPT to explain why:

'The word “than” is used in comparative structures, while “that” is used to introduce a clause, often a clause of indirect speech or a clause that gives extra information about something. In this case, “that” is introducing extra information about the Mahāvaṃsa statement, so it is the correct word to use.'